requirements.txt
mongomock>=3.22.1
pymongo>=3.10
pytest-asyncio>=0.11.0
pytest>=5.4
pyyaml>=5.1
